Move DDB Enhanced Client examples to their 'enhanced' subfolder
https://issues.amazon.com/issues/ded96dea-0cb3-4f2b-8708-5b47c5250df5
Update readme with new path
