--- javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/EnhancedBatchWriteItems.java
@@ -1,118 +1,0 @@
-//snippet-sourcedescription:[EnhancedBatchWriteItems.java demonstrates how to insert many items into an Amazon DynamoDB table by using the enhanced client.]
-//snippet-keyword:[SDK for Java v2]
-//snippet-service:[Amazon DynamoDB]
-
-/*
-   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   SPDX-License-Identifier: Apache-2.0
-*/
-package com.example.dynamodb;
-
-// snippet-start:[dynamodb.java2.mapping.batchitems.import]
-
-import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
-import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
-import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
-import software.amazon.awssdk.enhanced.dynamodb.Key;
-import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
-import software.amazon.awssdk.enhanced.dynamodb.model.BatchWriteItemEnhancedRequest;
-import software.amazon.awssdk.enhanced.dynamodb.model.WriteBatch;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
-import software.amazon.awssdk.services.dynamodb.model.DynamoDbException;
-
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-// snippet-end:[dynamodb.java2.mapping.batchitems.import]
-
-/*
- * Before running this code example, create an Amazon DynamoDB table named Customer with these columns:
- *   - id - the id of the record that is the key
- *   - custName - the customer name
- *   - email - the email value
- *   - registrationDate - an instant value when the item was added to the table
- *
- * Also, ensure that you have set up your development environment, including your credentials.
- *
- * For information, see this documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class EnhancedBatchWriteItems {
-
-    public static void main(String[] args) {
-
-        ProfileCredentialsProvider credentialsProvider = ProfileCredentialsProvider.create();
-        Region region = Region.US_EAST_1;
-        DynamoDbClient ddb = DynamoDbClient.builder()
-                .region(region)
-                .credentialsProvider(credentialsProvider)
-                .build();
-
-        DynamoDbEnhancedClient enhancedClient = DynamoDbEnhancedClient.builder()
-                .dynamoDbClient(ddb)
-                .build();
-
-        putBatchRecords(enhancedClient);
-        ddb.close();
-    }
-
-    // snippet-start:[dynamodb.java2.mapping.batchitems.main]
-    public static void putBatchRecords(DynamoDbEnhancedClient enhancedClient) {
-
-        try {
-            DynamoDbTable<Customer> customerMappedTable = enhancedClient.table("Customer", TableSchema.fromBean(Customer.class));
-            DynamoDbTable<Music> musicMappedTable = enhancedClient.table("Music", TableSchema.fromBean(Music.class));
-            LocalDate localDate = LocalDate.parse("2020-04-07");
-            LocalDateTime localDateTime = localDate.atStartOfDay();
-            Instant instant = localDateTime.toInstant(ZoneOffset.UTC);
-
-            Customer record2 = new Customer();
-            record2.setCustName("Fred Pink");
-            record2.setId("id110");
-            record2.setEmail("fredp@noserver.com");
-            record2.setRegistrationDate(instant) ;
-
-            Customer record3 = new Customer();
-            record3.setCustName("Susan Pink");
-            record3.setId("id120");
-            record3.setEmail("spink@noserver.com");
-            record3.setRegistrationDate(instant) ;
-
-            Customer record4 = new Customer();
-            record4.setCustName("Jerry orange");
-            record4.setId("id101");
-            record4.setEmail("jorange@noserver.com");
-            record4.setRegistrationDate(instant) ;
-
-
-            BatchWriteItemEnhancedRequest batchWriteItemEnhancedRequest = BatchWriteItemEnhancedRequest.builder()
-                    .writeBatches(
-                            WriteBatch.builder(Customer.class)          // add items to the Customer table
-                                    .mappedTableResource(customerMappedTable)
-                                    .addPutItem(builder -> builder.item(record2))
-                                    .addPutItem(builder -> builder.item(record3))
-                                    .addPutItem(builder -> builder.item(record4))
-                                    .build(),
-                            WriteBatch.builder(Music.class)             // delete an item from the Music table
-                                    .mappedTableResource(musicMappedTable)
-                                    .addDeleteItem(builder -> builder.key(
-                                            Key.builder().partitionValue("Famous Band").build()))
-                                    .build())
-                    .build();
-
-
-            // Add three items to the Customer table and delete one item from the Music table
-            enhancedClient.batchWriteItem(batchWriteItemEnhancedRequest);
-
-            System.out.println("done");
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-    // snippet-end:[dynamodb.java2.mapping.batchitems.main]
-}
