--- javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/EnhancedGetItem.java
@@ -1,80 +1,0 @@
-//snippet-sourcedescription:[EnhancedGetItem.java demonstrates how to retrieve an item from an Amazon DynamoDB table by using the enhanced client.]
-//snippet-keyword:[SDK for Java v2]
-//snippet-service:[Amazon DynamoDB]
-
-/*
-   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   SPDX-License-Identifier: Apache-2.0
-*/
-package com.example.dynamodb;
-
-// snippet-start:[dynamodb.java2.mapping.getitem.import]
-import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
-import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
-import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
-import software.amazon.awssdk.enhanced.dynamodb.Key;
-import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
-import software.amazon.awssdk.enhanced.dynamodb.model.GetItemEnhancedRequest;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
-import software.amazon.awssdk.services.dynamodb.model.DynamoDbException;
-// snippet-end:[dynamodb.java2.mapping.getitem.import]
-
-/*
- * Before running this code example, create an Amazon DynamoDB table named Customer with these columns:
- *   - id - the id of the record that is the key. Be sure one of the id values is `id101`
- *   - custName - the customer name
- *   - email - the email value
- *   - registrationDate - an instant value when the item was added to the table. These values
- *                        need to be in the form of `YYYY-MM-DDTHH:mm:ssZ`, such as 2022-07-11T00:00:00Z
- *
- * Also, ensure that you have set up your development environment, including your credentials.
- *
- * For information, see this documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-
-public class EnhancedGetItem {
-
-    public static void main(String[] args) {
-
-        ProfileCredentialsProvider credentialsProvider = ProfileCredentialsProvider.create();
-        Region region = Region.US_EAST_1;
-        DynamoDbClient ddb = DynamoDbClient.builder()
-            .credentialsProvider(credentialsProvider)
-            .region(region)
-            .build();
-
-        DynamoDbEnhancedClient enhancedClient = DynamoDbEnhancedClient.builder()
-            .dynamoDbClient(ddb)
-            .build();
-
-        getItem(enhancedClient);
-        ddb.close();
-    }
-
-    // snippet-start:[dynamodb.java2.mapping.getitem.main]
-    public static String getItem(DynamoDbEnhancedClient enhancedClient) {
-
-        Customer result = null;
-
-        try {
-            DynamoDbTable<Customer> table = enhancedClient.table("Customer", TableSchema.fromBean(Customer.class));
-            Key key = Key.builder()
-                .partitionValue("id101").sortValue("tred@noserver.com")
-                .build();
-
-            // Get the item by using the key.
-            result = table.getItem(
-                    (GetItemEnhancedRequest.Builder requestBuilder) -> requestBuilder.key(key));
-            System.out.println("******* The description value is " + result.getCustName());
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-        return result.getCustName();
-    }
-    // snippet-end:[dynamodb.java2.mapping.getitem.main]
-}
