--- javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/EnhancedPutItem.java
@@ -1,85 +1,0 @@
-//snippet-sourcedescription:[EnhancedPutItem.java demonstrates how to put an item into an Amazon DynamoDB table by using the enhanced client.]
-//snippet-keyword:[SDK for Java v2]
-//snippet-service:[Amazon DynamoDB]
-
-/*
-   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-   SPDX-License-Identifier: Apache-2.0
-*/
-
-package com.example.dynamodb;
-
-// snippet-start:[dynamodb.java2.mapping.putitem.import]
-import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
-import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
-import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
-import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
-import software.amazon.awssdk.services.dynamodb.model.DynamoDbException;
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-// snippet-end:[dynamodb.java2.mapping.putitem.import]
-
-/*
- * Before running this code example, create an Amazon DynamoDB table named Customer with these columns:
- *   - id - the id of the record that is the key
- *   - custName - the customer name
- *   - email - the email value
- *   - registrationDate - an instant value when the item was added to the table
- *
- * Also, ensure that you have set up your development environment, including your credentials.
- *
- * For information, see this documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class EnhancedPutItem {
-
-    public static void main(String[] args) {
-
-        ProfileCredentialsProvider credentialsProvider = ProfileCredentialsProvider.create();
-        Region region = Region.US_EAST_1;
-        DynamoDbClient ddb = DynamoDbClient.builder()
-            .credentialsProvider(credentialsProvider)
-            .region(region)
-            .build();
-
-        DynamoDbEnhancedClient enhancedClient = DynamoDbEnhancedClient.builder()
-            .dynamoDbClient(ddb)
-            .build();
-
-        putRecord(enhancedClient) ;
-        ddb.close();
-    }
-
-    // snippet-start:[dynamodb.java2.mapping.putitem.main]
-    public static void putRecord(DynamoDbEnhancedClient enhancedClient) {
-        try {
-            DynamoDbTable<Customer> custTable = enhancedClient.table("Customer", TableSchema.fromBean(Customer.class));
-
-            // Create an Instant value.
-            LocalDate localDate = LocalDate.parse("2020-04-07");
-            LocalDateTime localDateTime = localDate.atStartOfDay();
-            Instant instant = localDateTime.toInstant(ZoneOffset.UTC);
-
-            // Populate the Table.
-            Customer custRecord = new Customer();
-            custRecord.setCustName("Tom red");
-            custRecord.setId("id101");
-            custRecord.setEmail("tred@noserver.com");
-            custRecord.setRegistrationDate(instant) ;
-
-            // Put the customer data into an Amazon DynamoDB table.
-            custTable.putItem(custRecord);
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-        System.out.println("Customer data added to the table with id id101");
-    }
-    // snippet-end:[dynamodb.java2.mapping.putitem.main]
-}
