--- javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/EnhancedQueryRecordsWithSortKey.java
@@ -1,22 +1,0 @@
-package com.example.dynamodb;
-
-// snippet-start:[dynamodb.java2.mapping.querykey.import]
-// snippet-end:[dynamodb.java2.mapping.querykey.import]
-
-/*
- *This example is depreciated, see the Scenario example.
- */
-public class EnhancedQueryRecordsWithSortKey {
-
-    public static void main(String[] args) {
-
-        System.out.println("This example is depreciated, see the Scenario example");
-    }
-
-    // snippet-start:[dynamodb.java2.mapping.querykey.main]
-    public static void queryTableSort() {
-
-        System.out.println("This example is depreciated, see the Scenario example");
-    }
-    // snippet-end:[dynamodb.java2.mapping.querykey.main]
-}
