--- javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/EnhancedScanRecordsWithExpression.java
@@ -1,24 +1,0 @@
-package com.example.dynamodb;
-
-// snippet-start:[dynamodb.java2.mapping.scanEx.import]
-// snippet-end:[dynamodb.java2.mapping.scanEx.import]
-
-/**
- * This example is depreciated, see the Scenario example.
- */
-
-public class EnhancedScanRecordsWithExpression {
-
-    public static void main(String[] args) {
-
-        System.out.println("This example is depreciated, see the Scenario example");
-    }
-
-    // snippet-start:[dynamodb.java2.mapping.scanEx.main]
-    // Scan the table and retrieve only items where createDate is 2013-11-15.
-    public static void scanIndex() {
-
-        System.out.println("This example is depreciated, see the Scenario example");
-    }
-    // snippet-end:[dynamodb.java2.mapping.scanEx.main]
-}
