+++ javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/enhanced/EnhancedQueryRecords.java
@@ -1,0 +1,87 @@
+//snippet-sourcedescription:[EnhancedQueryRecords.java demonstrates how to query an Amazon DynamoDB table by using the enhanced client.]
+//snippet-keyword:[SDK for Java v2]
+//snippet-service:[Amazon DynamoDB]
+
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+package com.example.dynamodb.enhanced;
+
+// snippet-start:[dynamodb.java2.mapping.query.import]
+
+import com.example.dynamodb.Customer;
+import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
+import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
+import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
+import software.amazon.awssdk.enhanced.dynamodb.Key;
+import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
+import software.amazon.awssdk.enhanced.dynamodb.model.QueryConditional;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
+import software.amazon.awssdk.services.dynamodb.model.DynamoDbException;
+import java.util.Iterator;
+// snippet-end:[dynamodb.java2.mapping.query.import]
+
+/*
+ * Before running this code example, create an Amazon DynamoDB table named Customer with these columns:
+ *   - id - the id of the record that is the key
+ *   - custName - the customer name
+ *   - email - the email value
+ *   - registrationDate - an instant value when the item was added to the table
+ *
+ * Also, ensure that you have set up your development environment, including your credentials.
+ *
+ * For information, see this documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ */
+
+public class EnhancedQueryRecords {
+
+    public static void main(String[] args) {
+
+        ProfileCredentialsProvider credentialsProvider = ProfileCredentialsProvider.create();
+        Region region = Region.US_EAST_1;
+        DynamoDbClient ddb = DynamoDbClient.builder()
+            .region(region)
+            .credentialsProvider(credentialsProvider)
+            .build();
+
+        DynamoDbEnhancedClient enhancedClient = DynamoDbEnhancedClient.builder()
+            .dynamoDbClient(ddb)
+            .build();
+
+        String result = queryTable(enhancedClient);
+        System.out.println(result);
+        ddb.close();
+    }
+
+    // snippet-start:[dynamodb.java2.mapping.query.main]
+    public static String queryTable(DynamoDbEnhancedClient enhancedClient) {
+
+        try{
+            DynamoDbTable<Customer> mappedTable = enhancedClient.table("Customer", TableSchema.fromBean(Customer.class));
+            QueryConditional queryConditional = QueryConditional.keyEqualTo(Key.builder()
+                .partitionValue("id101")
+                .build());
+
+            // Get items in the table and write out the ID value.
+            Iterator<Customer> results = mappedTable.query(queryConditional).items().iterator();
+            String result="";
+
+            while (results.hasNext()) {
+                Customer rec = results.next();
+                result = rec.getId();
+                System.out.println("The record id is "+result);
+            }
+            return result;
+
+        } catch (DynamoDbException e) {
+            System.err.println(e.getMessage());
+            System.exit(1);
+        }
+        return "";
+    }
+    // snippet-end:[dynamodb.java2.mapping.query.main]
+ }
