+++ javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/enhanced/EnhancedTableSchema.java
@@ -1,0 +1,154 @@
+//snippet-sourcedescription:[EnhancedTableSchema.java demonstrates how to use the Amazon DynamoDB enhanced client and a TableSchema to put an item into a table.]
+//snippet-keyword:[SDK for Java v2]
+//snippet-service:[Amazon DynamoDB]
+
+/*
+   Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+   SPDX-License-Identifier: Apache-2.0
+*/
+package com.example.dynamodb.enhanced;
+
+// snippet-start:[dynamodb.java2.mapping.tableschema.import]
+import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
+import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
+import software.amazon.awssdk.enhanced.dynamodb.DynamoDbTable;
+import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
+import software.amazon.awssdk.enhanced.dynamodb.mapper.StaticTableSchema;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
+import software.amazon.awssdk.services.dynamodb.model.DynamoDbException;
+import static software.amazon.awssdk.enhanced.dynamodb.mapper.StaticAttributeTags.primaryPartitionKey;
+import static software.amazon.awssdk.enhanced.dynamodb.mapper.StaticAttributeTags.primarySortKey;
+import static software.amazon.awssdk.enhanced.dynamodb.mapper.StaticAttributeTags.secondaryPartitionKey;
+import static software.amazon.awssdk.enhanced.dynamodb.mapper.StaticAttributeTags.secondarySortKey;
+// snippet-end:[dynamodb.java2.mapping.tableschema.import]
+
+/*
+ * Before running this code example, set up your development environment, including your credentials.
+ *
+ * For information, see this documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ */
+public class EnhancedTableSchema {
+
+    private static final TableSchema<Record> TABLE_SCHEMA =
+            StaticTableSchema.builder(Record.class)
+                    .newItemSupplier(Record::new)
+                    .addAttribute(String.class, a -> a.name("id")
+                            .getter(Record::getId)
+                            .setter(Record::setId)
+                            .tags(primaryPartitionKey()))
+                    .addAttribute(String.class, a -> a.name("sort")
+                            .getter(Record::getSort)
+                            .setter(Record::setSort)
+                            .tags(primarySortKey()))
+                    .addAttribute(String.class, a -> a.name("attribute")
+                            .getter(Record::getAttribute)
+                            .setter(Record::setAttribute))
+                    .addAttribute(String.class, a -> a.name("attribute2*")
+                            .getter(Record::getAttribute2)
+                            .setter(Record::setAttribute2)
+                            .tags(secondaryPartitionKey("gsi_1")))
+                    .addAttribute(String.class, a -> a.name("attribute3")
+                            .getter(Record::getAttribute3)
+                            .setter(Record::setAttribute3)
+                            .tags(secondarySortKey("gsi_1")))
+                    .build();
+
+    public static void main(String[] args) {
+
+        ProfileCredentialsProvider credentialsProvider = ProfileCredentialsProvider.create();
+        Region region = Region.US_EAST_1;
+        DynamoDbClient ddb = DynamoDbClient.builder()
+                .credentialsProvider(credentialsProvider)
+                .region(region)
+                .build();
+
+        DynamoDbEnhancedClient enhancedClient = DynamoDbEnhancedClient.builder()
+                .dynamoDbClient(ddb)
+                .build();
+
+        putRecord(enhancedClient);
+        ddb.close();
+    }
+
+    // snippet-start:[dynamodb.java2.mapping.tableschema.main]
+    public static void putRecord(DynamoDbEnhancedClient enhancedClient){
+
+        try {
+            //Create a DynamoDbTable object
+            DynamoDbTable<Record> mappedTable = enhancedClient.table("Record", TABLE_SCHEMA);
+
+            //Populate the Table
+            Record record = new Record()
+                    .setId("id-value")
+                    .setSort("sort-value")
+                    .setAttribute("one")
+                    .setAttribute2("two")
+                    .setAttribute3("three");
+
+            mappedTable.putItem(record);
+
+        } catch (DynamoDbException e) {
+            System.err.println(e.getMessage());
+            System.exit(1);
+        }
+        System.out.println("done");
+    }
+    // snippet-end:[dynamodb.java2.mapping.tableschema.main]
+
+    // Define the Record class that is used to map to the DynamoDB table
+    private static class Record {
+        private String id;
+        private String sort;
+        private String attribute;
+        private String attribute2;
+        private String attribute3;
+
+        private String getId() {
+            return id;
+        }
+
+        private Record setId(String id) {
+            this.id = id;
+            return this;
+        }
+
+        private String getSort() {
+            return sort;
+        }
+
+        private Record setSort(String sort) {
+            this.sort = sort;
+            return this;
+        }
+
+        private String getAttribute() {
+            return attribute;
+        }
+
+        private Record setAttribute(String attribute) {
+            this.attribute = attribute;
+            return this;
+        }
+
+        private String getAttribute2() {
+            return attribute2;
+        }
+
+        private Record setAttribute2(String attribute2) {
+            this.attribute2 = attribute2;
+            return this;
+        }
+
+        private String getAttribute3() {
+            return attribute3;
+        }
+
+        private Record setAttribute3(String attribute3) {
+            this.attribute3 = attribute3;
+            return this;
+        }
+    }
+}
