+++ javav2/example_code/dynamodb/src/test/java/EnhancedClientTest.java
@@ -1,6 +1,7 @@
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.example.dynamodb.*;
+import com.example.dynamodb.enhanced.*;
 import software.amazon.awssdk.enhanced.dynamodb.DynamoDbEnhancedClient;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
